Extract common logic from ApplicationAssetCache and AssetFactory into common base class (fixes absolute web css references) [CP b591052a81e50693e935e87d72bc8f94fc85fc5f from ol4]
