+++ src/Mapbender/CoreBundle/Asset/ApplicationAssetCache.php
@@ -3,7 +3,6 @@
 namespace Mapbender\CoreBundle\Asset;
 
 use Assetic\Asset\AssetCollection;
-use Assetic\Asset\FileAsset;
 use Assetic\Asset\StringAsset;
 use Assetic\AssetManager;
 use Assetic\Cache\FilesystemCache;
@@ -16,11 +15,8 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * @author Andriy Oblivantsev <andriy.oblivantsev@wheregroup.com>
  * @package Mapbender\CoreBundle\Asset
  */
-class ApplicationAssetCache
+class ApplicationAssetCache extends AssetFactoryBase
 {
-    /** @var ContainerInterface  */
-    protected $container;
-
     /** @var array|\Assetic\Asset\FileAsset[]|\Assetic\Asset\StringAsset[]  */
     protected $inputs;
 
@@ -42,7 +38,7 @@ class ApplicationAssetCache
      */
     public function __construct(ContainerInterface $container, $inputs, $type, $force = false)
     {
-        $this->container = $container;
+        parent::__construct($container);
         $this->inputs = $inputs;
         $this->type = $type;
         $this->targetPath = $this->getTargetPath();
@@ -60,12 +56,10 @@ class ApplicationAssetCache
 
         // For each asset build compiled, cached asset
         $assetRootPath = $this->getWebDir();
-        $assetTargetPath = $this->targetPath;
 
         $assets = new AssetCollection(array(), array(), $assetRootPath);
         $assets->setTargetPath($this->targetPath);
 
-        $locator = $this->container->get('file_locator');
         $manager = new AssetManager();
         $cache = new FilesystemCache($static_assets_cache_path);
         $devCache = new FilesystemCache($static_assets_cache_path . '/.dev-cache');
@@ -78,25 +72,8 @@ class ApplicationAssetCache
                 $manager->set($name, $input);
                 continue;
             }
+            $fileAsset = $this->makeFileAsset($input, $this->type);
 
-            // First, build file asset with filters and public path information
-            $sourcePath = $this->getSourcePath($input);
-            if ($sourcePath) {
-                $file = $locator->locate($sourcePath);
-            } else {
-                $file = $locator->locate($input);
-            }
-
-            // Build filter list (None for JS/Trans, Compass for SASS and Rewrite for SASS/CSS)
-            $filters = array();
-            if('css' === $this->type) {
-                if('scss' === pathinfo($file, PATHINFO_EXTENSION)) {
-                    $filters[] = $this->container->get('assetic.filter.compass');
-                }
-                $filters[] = $this->container->get('assetic.filter.cssrewrite');
-            }
-
-            $fileAsset = new FileAsset($file, $filters, null, $assetRootPath . '/' . $sourcePath);
             $fileAsset->setTargetPath($this->targetPath);
 
             $name = str_replace(array('@', 'Resources/public/'), '', $input);
@@ -112,7 +89,7 @@ class ApplicationAssetCache
             $needsCompiling = false;
 
             // SASS things need to be compiled.
-            if('scss' === pathinfo($file, PATHINFO_EXTENSION)) $needsCompiling = true;
+            if('scss' === pathinfo($fileAsset->getSourcePath(), PATHINFO_EXTENSION)) $needsCompiling = true;
 
             if($needsCompiling) {
                 if($isDevPlus) {
@@ -144,42 +121,28 @@ class ApplicationAssetCache
     /**
      * @return string
      */
-    protected function getWebDir()
-    {
-        return dirname($this->container->getParameter('kernel.root_dir')) . '/web';
-    }
-
-    /**
-     * @param $input
-     * @return string
-     */
-    protected function getSourcePath($input)
-    {
-        if ($input[0] == '@') {
-            // Bundle name
-            $bundle = substr($input, 1, strpos($input, '/') - 1);
-            // Path inside the Resources/public folder
-            $assetPath = substr($input,
-                strlen('@' . $bundle . '/Resources/public'));
-            $assetDir = 'bundles/' . preg_replace('/bundle$/', '', strtolower($bundle));
-
-            return $this->getSourcePath($assetDir . $assetPath);
-        } else {
-            $webRoot = $this->getWebDir();
-            $inWeb = $webRoot . '/' . ltrim($input, '/');
-            if (@is_file($inWeb) && @is_readable($inWeb)) {
-                return $inWeb;
-            }
-        }
-    }
-
-    /**
-     * @return string
-     */
     protected function getTargetPath()
     {
         $route = $this->container->get('router')->getRouteCollection()->get('mapbender_core_application_assets');
         $target = str_replace('\\', '/', realpath($this->container->get('kernel')->getRootDir() . '/../web/app.php')) . $route->getPath();
         return $target;
+    }
+
+    /**
+     * @param string $fileName
+     * @param string $assetType one of 'js', 'css', 'trans'
+     * @return object[]
+     * @todo: figure out assetic filter base class
+     */
+    protected function getFilters($fileName, $assetType)
+    {
+        $filters = array();
+        if ('css' === $assetType) {
+            if ('scss' === pathinfo($fileName, PATHINFO_EXTENSION)) {
+                $filters[] = $this->container->get('assetic.filter.compass');
+            }
+            $filters[] = $this->container->get('assetic.filter.cssrewrite');
+        }
+        return $filters;
     }
 }
