+++ src/Mapbender/CoreBundle/Asset/AssetFactory.php
@@ -14,11 +14,8 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * @author Andriy Oblivantsev <andriy.oblivantsev@wheregroup.com>
  * @package Mapbender\CoreBundle\Asset
  */
-class AssetFactory
+class AssetFactory extends AssetFactoryBase
 {
-    /** @var ContainerInterface  */
-    protected $container;
-
     /** @var array|\Assetic\Asset\FileAsset[]|\Assetic\Asset\StringAsset[]  */
     protected $inputs;
 
@@ -45,6 +42,7 @@ class AssetFactory
      */
     public function __construct(ContainerInterface $container, array $inputs, $type, $targetPath, $sourcePath)
     {
+        parent::__construct($container);
         $this->sourcePath = $sourcePath;
         $this->container  = $container;
         $this->inputs     = $inputs;
@@ -61,32 +59,21 @@ class AssetFactory
             $assetRootPath = dirname($this->container->getParameter('kernel.root_dir')) . '/web';
             $this->collection = new AssetCollection(array(), array(), $assetRootPath);
             $this->collection->setTargetPath($this->targetPath);
-            $locator = $this->container->get('file_locator');
             $manager = new AssetManager();
             $stringAssetCounter = 0;
 
             foreach($this->inputs as $input) {
                 // GUI declared CSS
-                if($input instanceof StringAsset) {
+                if ($input instanceof StringAsset) {
                     $name = 'stringasset_' . $stringAssetCounter++;
                     $manager->set($name, $input);
-                    continue;
+                } else {
+                    $fileAsset = $this->makeFileAsset($input, $this->type);
+                    $fileAsset->setTargetPath($this->targetPath);
+                    $name = str_replace(array('@', 'Resources/public/'), '', $input);
+                    $name = str_replace(array('/', '.', '-'), '__', $name);
+                    $manager->set($name, $fileAsset);
                 }
-
-                // First, build file asset with public path information
-                $file = $locator->locate($input);
-                $publicSourcePath = $assetRootPath . '/' . $this->getPublicSourcePath($input);
-
-                $fileAsset = new FileAsset(
-                    $file,
-                    array(),
-                    $assetRootPath,
-                    $publicSourcePath);
-                $fileAsset->setTargetPath($this->targetPath);
-
-                $name = str_replace(array('@', 'Resources/public/'), '', $input);
-                $name = str_replace(array('/', '.', '-'), '__', $name);
-                $manager->set($name, $fileAsset);
             }
 
             // Finally, wrap everything into a single asset collection
@@ -138,20 +125,4 @@ class AssetFactory
         return implode($imports, "\n") . "\n" . $content;
     }
 
-    /**
-     * @param string $input
-     * @return string
-     */
-    protected function getPublicSourcePath($input)
-    {
-        $sourcePath = null;
-        if ($input[0] == '@') {
-            // Bundle name
-            $bundle = substr($input, 1, strpos($input, '/') - 1);
-            // Path inside the Resources/public folder
-            $assetPath = substr($input, strlen('@' . $bundle . '/Resources/public'));
-
-            return 'bundles/' . preg_replace('/bundle$/', '', strtolower($bundle)) . $assetPath;
-        }
-    }
 }
